Move MainActivity to ui directory
