+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
             android:value="${MAPS_API_KEY}" />
 
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.ArtChart.NoActionBar">
             <intent-filter>
