+++ app/src/main/java/com/lab/artchart/ui/search/SearchFragment.kt
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.lab.artchart.MainActivity
+import com.lab.artchart.ui.MainActivity
 import com.lab.artchart.databinding.FragmentSearchBinding
 
 class SearchFragment : Fragment() {
